Fix nouveau warning when returning pointers in uint64_t objects.

+++ shared-core/nouveau_state.c
@@ -614,7 +614,7 @@ int nouveau_ioctl_getparam(struct drm_device *dev, void *data, struct drm_file *
 		break;
 	case NOUVEAU_GETPARAM_PCI_PHYSICAL:
 		if ( dev -> sg )
-			getparam->value=(uint64_t) dev->sg->virtual;
+			getparam->value=(uint64_t) (uintptr_t) dev->sg->virtual;
 		else
 		     {
 		     DRM_ERROR("Requested PCIGART address, while no PCIGART was created\n");
